Split up source into multiple files
