+++ src/cursor.h
@@ -1,0 +1,9 @@
+#include "server.h"
+#include "wayland.h"
+
+#ifndef _CURSOR_H_
+#define _CURSOR_H_
+
+void cursor_init(Server *);
+
+#endif
